backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
SQLAlchemy==2.0.43
aiosqlite==0.21.0
PyJWT==2.10.1
